Move tests from DescriptorSet to DescSet
