+++ packages/protobuf-test/src/next/helpers.ts
@@ -46,18 +46,25 @@ export function testGenerated<Desc extends DescMessage>(
   });
 }
 
+let upstreamProtobuf: UpstreamProtobuf | undefined;
+
+export async function compileFileDescriptorSet(
+  files: Record<string, string>,
+): Promise<FileDescriptorSet> {
+  upstreamProtobuf = upstreamProtobuf ?? new UpstreamProtobuf();
+  const bytes = await upstreamProtobuf.compileToDescriptorSet(files, {
+    includeImports: true,
+    retainOptions: true,
+  });
+  return FileDescriptorSet.fromBinary(bytes);
+}
+
 export async function compileMessage(proto: string): Promise<DescMessage> {
-  const upstream = new UpstreamProtobuf();
-  const setBin = await upstream.compileToDescriptorSet(
-    {
+  const set = createDescFileSet(
+    await compileFileDescriptorSet({
       "input.proto": proto,
-    },
-    {
-      includeImports: true,
-      retainOptions: true,
-    },
+    }),
   );
-  const set = createDescFileSet(FileDescriptorSet.fromBinary(setBin));
   const file = set.getFile("input.proto");
   if (file === undefined) {
     throw new Error("missing file descriptor for input.proto");
